Assistant checkpoint: Updated home and donation form designs
Assistant generated file changes:
- locations/templates/locations/home.html: Update home page design
- locations/templates/locations/donation_form.html: Update donation form design

---

User prompt:

Act as an experienced product designer.
I would like to design a market place for donations.The website is dedicated for Israelis and should be placed in heberw.
The website will have 3 pages.
1. Home page
2. Form page
3. Results page
The design should be highly intuitive and friendly to use.

The home should provide some general informations about what is this website is about, the idea is to make donations simple and easy. we connect between good people with the intention to donate some of their personal belongingss.

The form is where we ask 3 questions
What is the user trying to donate?
How far is he willing to travel in order to make the donations?
Whats is home addrees?

The results page should provide the list of locations to donate ordered by the disatnce from his home address (provide before)